refactor: refactoring redux slices
