+++ view/desktop/src/store/languages/index.ts
@@ -1,0 +1,1 @@
+export { default as languagesReducer } from "./languagesSlice";
