+++ view/desktop/src/store/themes/index.ts
@@ -1,0 +1,2 @@
+export { default as themesReducer } from "./themesSlice";
+export * from "./themesThunks";
